move flask and database to separate files
